Update codebase to make it more OOP
Codebase has level-7 requirements.

Most of the logic still runs in the main class.

Adding the following features allows us to make the program more OOP
* refactor code in the main class into different classes
