--- src/main/java/DukeException.java
@@ -1,2 +1,0 @@
-public class DukeException extends Exception {
-}
